Fix for issue 47
There was a severe bug in the simulator where comparison was being made
between the timed notice and the notice which has been just
concluded. But the way the code is the concluded notice was getting
freed-up (as part of a freelist) and therefor the comparison was always
returning false.