Purged all NOP constraints stuff
After refactoring there is a much cleaner way to achieve wat needs to
be done. Basically we need to look for all the dependencies and if a
dependency is not a var for any of the constraints and therefor is a
pure dependency only, we need to randomize it first and then flag it
as solved.